Move EntityCollisionResult to its own class (#2345)
* Move EntityCollisionResult to its own class
